permissions addes

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -4,6 +4,7 @@ const { randomBytes } = require("crypto");
 const { promisify } = require("util");
 
 const { transport, makeANiceEmail } = require("../mail");
+const { hasPermission } = require("../utils");
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
@@ -41,7 +42,15 @@ const Mutations = {
   },
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
-    const item = await ctx.db.query.item({ where }, `{id title}`);
+    const item = await ctx.db.query.item({ where }, `{id title user {id}}`);
+    // Check if we have the permission
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ["ADMIN", "ITEMDELETE"].includes(permission)
+    );
+    if (!ownsItem && hasPermissions) {
+      throw new Error("You don't have permission to do that");
+    }
     return ctx.db.mutation.deleteItem({ where });
   },
   async signup(parent, args, ctx, info) {
@@ -148,6 +157,31 @@ const Mutations = {
     });
     // return new user
     return updateUser;
+  },
+  async updatePermissions(parent, args, ctx, info) {
+    // Check logged in
+    if (!ctx.request.userId) {
+      throw new Error("You must be logged in!");
+    }
+    // query the user
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.request.userId
+        }
+      },
+      info
+    );
+    // check
+    hasPermission(currentUser, ["ADMIN", "PERMISSIONUPDATE"]);
+    // update the user
+    return ctx.db.mutation.updateUser(
+      {
+        data: { permissions: { set: args.permissions } },
+        where: { id: args.userId }
+      },
+      info
+    );
   }
 };
